# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateCaseRequest": {
        "properties": {
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "risk_score_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score Id"
          }
        },
        "required": [
          "customer_id"
        ],
        "title": "CreateCaseRequest",
        "type": "object"
      },
      "CreateMissionRequest": {
        "properties": {
          "assigned_team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Team Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "scheduled_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateMissionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InspectionCaseOut": {
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mission_id": {
            "title": "Mission Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "priority_rank": {
            "title": "Priority Rank",
            "type": "integer"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "risk_score_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score Id"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "mission_id",
          "customer_id",
          "risk_score_id",
          "status",
          "priority_rank",
          "scheduled_at",
          "assigned_to",
          "recommended_action",
          "notes"
        ],
        "title": "InspectionCaseOut",
        "type": "object"
      },
      "InspectionMissionOut": {
        "properties": {
          "assigned_team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Team Id"
          },
          "case_count": {
            "title": "Case Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "estimated_total_value": {
            "title": "Estimated Total Value",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "route_geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Geometry"
          },
          "scheduled_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "region_id",
          "status",
          "scheduled_date",
          "assigned_team_id",
          "route_geometry",
          "estimated_total_value",
          "currency",
          "created_by",
          "created_at",
          "case_count"
        ],
        "title": "InspectionMissionOut",
        "type": "object"
      },
      "InspectionOutcomeOut": {
        "properties": {
          "confirmed_loss_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed Loss Type"
          },
          "estimated_recovered_kwh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Recovered Kwh"
          },
          "estimated_recovered_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Recovered Value"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inspection_case_id": {
            "title": "Inspection Case Id",
            "type": "string"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "submitted_by": {
            "title": "Submitted By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "inspection_case_id",
          "outcome",
          "confirmed_loss_type",
          "estimated_recovered_kwh",
          "estimated_recovered_value",
          "evidence",
          "submitted_by",
          "submitted_at"
        ],
        "title": "InspectionOutcomeOut",
        "type": "object"
      },
      "InspectionQueueItem": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "estimated_loss_kwh": {
            "title": "Estimated Loss Kwh",
            "type": "number"
          },
          "estimated_loss_value": {
            "title": "Estimated Loss Value",
            "type": "number"
          },
          "external_ref": {
            "title": "External Ref",
            "type": "string"
          },
          "inspection_priority": {
            "title": "Inspection Priority",
            "type": "number"
          },
          "recommended_action": {
            "title": "Recommended Action",
            "type": "string"
          },
          "region_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Name"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "number"
          },
          "risk_tier": {
            "title": "Risk Tier",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "external_ref",
          "risk_score",
          "risk_tier",
          "inspection_priority",
          "estimated_loss_kwh",
          "estimated_loss_value",
          "currency",
          "recommended_action",
          "region_name"
        ],
        "title": "InspectionQueueItem",
        "type": "object"
      },
...
      "Page_InspectionMissionOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InspectionMissionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "Page[InspectionMissionOut]",
        "type": "object"
      },
      "Page_InspectionQueueItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InspectionQueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "Page[InspectionQueueItem]",
        "type": "object"
      },
...
      "RouteRequest": {
        "properties": {
          "customer_ids": {
            "items": {
              "type": "string"
            },
            "title": "Customer Ids",
            "type": "array"
          }
        },
        "required": [
          "customer_ids"
        ],
        "title": "RouteRequest",
        "type": "object"
      },
      "RouteResponse": {
        "properties": {
          "ordered_customer_ids": {
            "items": {
              "type": "string"
            },
            "title": "Ordered Customer Ids",
            "type": "array"
          },
          "route_geometry": {
            "additionalProperties": true,
            "title": "Route Geometry",
            "type": "object"
          },
          "total_distance_km": {
            "title": "Total Distance Km",
            "type": "number"
          }
        },
        "required": [
          "ordered_customer_ids",
          "route_geometry",
          "total_distance_km"
        ],
        "title": "RouteResponse",
        "type": "object"
      },
      "SubmitOutcomeRequest": {
        "properties": {
          "confirmed_loss_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed Loss Type"
          },
          "estimated_recovered_kwh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Recovered Kwh"
          },
          "estimated_recovered_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Recovered Value"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "SubmitOutcomeRequest",
        "type": "object"
      },
...
      "UpdateCaseRequest": {
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "UpdateCaseRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Authoritative operational API for non-technical loss detection.",
    "title": "GridTrace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inspections/cases/{case_id}": {
      "patch": {
        "operationId": "update_case_api_v1_inspections_cases__case_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionCaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Case",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/cases/{case_id}/outcome": {
      "post": {
        "operationId": "submit_outcome_api_v1_inspections_cases__case_id__outcome_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitOutcomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionOutcomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Outcome",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/missions": {
      "get": {
        "operationId": "list_missions_api_v1_inspections_missions_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InspectionMissionOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Missions",
        "tags": [
          "inspections"
        ]
      },
      "post": {
        "operationId": "create_mission_api_v1_inspections_missions_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionMissionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Mission",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/missions/{mission_id}/cases": {
      "post": {
        "operationId": "add_case_api_v1_inspections_missions__mission_id__cases_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionCaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Case",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/queue": {
      "get": {
        "operationId": "queue_api_v1_inspections_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "region_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Id"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InspectionQueueItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/route": {
      "post": {
        "operationId": "build_route_api_v1_inspections_route_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Route",
        "tags": [
          "inspections"
        ]
      }
    },
...
  }
}